Uso mayusculas y reorganización carpetas
