+++ src/controllers/users/userControllers.js
@@ -1,27 +1,27 @@
-import Users from '../models/userModels.js'
+import Users from '../../models/userModels.js'
 
 //Create user
-export const createusers = async (request, response) => {
-  const {name, lastname, email, rol} = request.body;
-  const newUser= new Users({name, lastname, email, rol});
+export const createUsers = async (request, response) => {
+  const {name, lastName, email, rol} = request.body;
+  const newUser= new Users({name, lastName, email, rol});
   const userSaved = await newUser.save();
-  response.json(userSaved);
+  response.status(200).json(userSaved);
 }
 
 //Get all users
-export const getusers = async (request, response) => {
+export const getUsers = async (request, response) => {
   const userGet = await Users.find();
   response.json(userGet);
 }
 
 //Get user
-export const getusersById = async (request, response) => {
+export const getUsersById = async (request, response) => {
   const userGetById = await Users.findById(request.params.usersId);
   response.json(userGetById);
 }
 
 //Update user
-export const updateusersById = async (request, response) => {
+export const updateUsersById = async (request, response) => {
   const userUpdate = await Users.findByIdAndUpdate(request.params.usersId, request.body, {
     new: true
   });
@@ -29,7 +29,7 @@ export const updateusersById = async (request, response) => {
 }
 
 //Delete user
-export const deleteusersById = async (request, response) => {
+export const deleteUsersById = async (request, response) => {
   await Users.findByIdAndDelete(request.params.usersId);
   response.json("user delete");
 }
