+++ src/models/userModels.js
@@ -2,8 +2,8 @@ import { Schema, model} from "mongoose";
 
 const usersSchema = new Schema({
   name: String,
-  lastname: String,
-  email: String,
+  lastName: String,
+  email: { type: String, unique: true },
   rol: String
 }, {
   timestamps: true,
